fix issues caused by question ids no longer being unique between question types

+++ surveyTool/src/main/java/iks/surveytool/dtos/SurveyResponseDTO.java
@@ -12,7 +12,7 @@ import java.util.Map;
 public class SurveyResponseDTO extends AbstractDTO {
     private long surveyId;
     private Long userId;
-    private Map<Long, AnswerDTO> answers = new HashMap<>();
+    private Map<Integer, AnswerDTO> answers = new HashMap<>();
     
     @Override
     public void resetId() {
